flyttet knapper til footer i delemodal
