+++ frontend/mulighetsrommet-veileder-flate/src/apps/modia/delMedBruker/DelMedBrukerContent.tsx
@@ -1,18 +1,9 @@
-import {
-  Alert,
-  Button,
-  Checkbox,
-  ErrorMessage,
-  HStack,
-  HelpText,
-  Textarea,
-} from "@navikt/ds-react";
+import { Alert, ErrorMessage, Textarea } from "@navikt/ds-react";
 import { DelMedBruker, VeilederflateTiltaksgjennomforing } from "mulighetsrommet-api-client";
 import React, { Dispatch, useEffect, useRef } from "react";
 import { erPreview, formaterDato } from "@/utils/Utils";
-import delemodalStyles from "./Delemodal.module.scss";
+import styles from "./Delemodal.module.scss";
 import { Actions, State } from "./DelemodalActions";
-import { useLogEvent } from "@/logging/amplitude";
 import { getDelMedBrukerTekst } from "@/apps/modia/delMedBruker/helpers";
 
 export const MAKS_ANTALL_TEGN_DEL_MED_BRUKER = 500;
@@ -24,6 +15,7 @@ interface Props {
   brukernavn?: string;
   harDeltMedBruker?: DelMedBruker;
   tiltaksgjennomforing: VeilederflateTiltaksgjennomforing;
+  enableRedigerDeletekst: boolean;
 }
 
 export function DelMedBrukerContent({
@@ -33,12 +25,11 @@ export function DelMedBrukerContent({
   brukernavn,
   harDeltMedBruker,
   tiltaksgjennomforing,
+  enableRedigerDeletekst,
 }: Props) {
-  const { enableRedigerDeletekst } = state;
   const endreDeletekstRef = useRef<HTMLTextAreaElement>(null);
   const datoSidenSistDelt =
     harDeltMedBruker?.createdAt && formaterDato(new Date(harDeltMedBruker.createdAt));
-  const { logEvent } = useLogEvent();
 
   const standardtekstLengde = state.deletekst.length;
 
@@ -47,14 +38,6 @@ export function DelMedBrukerContent({
       endreDeletekstRef?.current?.focus();
     }
   }, [enableRedigerDeletekst]);
-
-  const enableEndreDeletekst = () => {
-    dispatch({ type: "Enable rediger deletekst", payload: true });
-    logEvent({
-      name: "arbeidsmarkedstiltak.del-med-bruker",
-      data: { action: "Endre deletekst", tiltakstype: tiltaksgjennomforing.tiltakstype.navn },
-    });
-  };
 
   const forMangeTegn = (tekst: string): boolean => {
     return tekst.length > standardtekstLengde + MAKS_ANTALL_TEGN_DEL_MED_BRUKER;
@@ -71,14 +54,18 @@ export function DelMedBrukerContent({
   return (
     <>
       {harDeltMedBruker ? (
-        <Alert variant="warning">{`Dette tiltaket ble delt med bruker ${datoSidenSistDelt}.`}</Alert>
+        <Alert
+          variant="warning"
+          className={styles.top_warning}
+        >{`Dette tiltaket ble delt med bruker ${datoSidenSistDelt}.`}</Alert>
       ) : null}
 
+      {/*<div className={styles.scroll}>*/}
       <Textarea
         label="Tekst som deles med bruker"
         hideLabel
         readOnly={!enableRedigerDeletekst}
-        className={delemodalStyles.deletekst}
+        className={styles.deletekst}
         error={handleError()}
         ref={endreDeletekstRef}
         size="medium"
@@ -89,60 +76,22 @@ export function DelMedBrukerContent({
       >
         {state.deletekst}
       </Textarea>
-
-      <HStack gap="4">
-        {enableRedigerDeletekst ? null : (
-          <Button
-            onClick={enableEndreDeletekst}
-            variant="secondary"
-            className={delemodalStyles.endreTekst_btn}
-          >
-            Rediger melding
-          </Button>
-        )}
-        <HStack gap="1" style={{ marginTop: "1rem" }}>
-          <Checkbox
-            onChange={(e) => {
-              dispatch({
-                type: "Venter på svar fra bruker",
-                payload: e.currentTarget.checked,
-              });
-              if (e.currentTarget.checked) {
-                logEvent({
-                  name: "arbeidsmarkedstiltak.del-med-bruker",
-                  data: {
-                    action: "Sett venter på svar fra bruker",
-                    tiltakstype: tiltaksgjennomforing.tiltakstype.navn,
-                  },
-                });
-              }
-            }}
-            checked={state.venterPaaSvarFraBruker}
-            value="venter-pa-svar-fra-bruker"
-          >
-            Venter på svar fra bruker
-          </Checkbox>
-          <HelpText title="Hva betyr dette valget?">
-            Ved å huke av for at du venter på svar fra bruker vil du kunne bruke filteret i
-            oversikten til å se alle brukere du venter på svar fra.
-          </HelpText>
-        </HStack>
-      </HStack>
+      {/*</div>*/}
 
       {!veiledernavn ? (
-        <ErrorMessage className={delemodalStyles.feilmeldinger}>
+        <ErrorMessage className={styles.feilmeldinger}>
           • Kunne ikke hente veileders navn
         </ErrorMessage>
       ) : null}
 
       {!brukernavn ? (
-        <ErrorMessage className={delemodalStyles.feilmeldinger}>
+        <ErrorMessage className={styles.feilmeldinger}>
           • Kunne ikke hente brukers navn
         </ErrorMessage>
       ) : null}
 
       {!getDelMedBrukerTekst(tiltaksgjennomforing) ? (
-        <ErrorMessage className={delemodalStyles.feilmeldinger}>
+        <ErrorMessage className={styles.feilmeldinger}>
           • Mangler ferdigutfylt tekst som kan deles med bruker{" "}
         </ErrorMessage>
       ) : null}
@@ -151,7 +100,7 @@ export function DelMedBrukerContent({
         <Alert
           variant="warning"
           data-testid="alert-preview-del-med-bruker"
-          className={delemodalStyles.preview_alert}
+          className={styles.preview_alert}
         >
           Det er ikke mulig å dele tiltak med bruker i forhåndsvisning. Brukers navn blir automatisk
           satt utenfor forhåndsvisningsmodus.
